refactor: Car 객체 생성의 책임을 가지는 객체 기존 Participants에서 Race로 수정
